Move from_key_and_merkle_root to extension trait
The from_key_and_merkle_root method on the TapTweakHash type is the
only part of the taproot-primitives crate that causes a dependency on
the crypto crate. By moving this back to bitcoin, we can remove the
dependency on crypto, allows these types to stabilise in primitives
without waiting on crypto to stabilise.

Move from_key_and_merkle_root to new extension trait TapTweakHashExt in
bitcoin.
