+++ bitcoin/src/taproot/mod.rs
@@ -99,10 +99,35 @@ crate::internal_macros::define_extension_trait! {
     }
 }
 
+crate::internal_macros::define_extension_trait! {
+    /// Extension functionality for the [`TapTweakHash`] type.
+    pub trait TapTweakHashExt impl for TapTweakHash {
+        /// Constructs a new BIP-0341 [`TapTweakHash`] from key and Merkle root. Produces `H_taptweak(P||R)` where
+        /// `P` is the internal key and `R` is the Merkle root.
+        fn from_key_and_merkle_root<K: Into<UntweakedPublicKey>>(
+            internal_key: K,
+            merkle_root: Option<TapNodeHash>,
+        ) -> Self {
+            let internal_key = internal_key.into();
+            let mut eng = sha256t::Hash::<TapTweakTag>::engine();
+            // always hash the key
+            eng.input(&internal_key.serialize().0);
+            if let Some(h) = merkle_root {
+                eng.input(h.as_ref());
+            } else {
+                // nothing to hash
+            }
+            let inner = sha256t::Hash::<TapTweakTag>::from_engine(eng);
+            Self::from_byte_array(inner.to_byte_array())
+        }
+    }
+}
+
 mod sealed {
     pub trait Sealed {}
     impl Sealed for super::TapLeafHash {}
     impl Sealed for super::TapNodeHash {}
+    impl Sealed for super::TapTweakHash {}
 }
 
 /// Computes branch hash given two hashes of the nodes underneath it and returns
